Split command interface and service modules
