+++ nym-vpnd/src/cli.rs
@@ -1,3 +1,6 @@
+// Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
 use clap::Parser;
 use nym_vpn_lib::nym_bin_common::bin_info_local_vergen;
 use std::{path::PathBuf, sync::OnceLock};
@@ -26,4 +29,3 @@ fn check_path(path: &str) -> Result<PathBuf, String> {
     }
     Ok(path)
 }
-
