+++ nym-vpnd/src/service/mod.rs
@@ -1,0 +1,12 @@
+// Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
+mod exit_listener;
+mod service;
+mod start;
+mod status_listener;
+
+pub(crate) use service::{
+    VpnServiceCommand, VpnServiceConnectResult, VpnServiceDisconnectResult, VpnServiceStatusResult,
+};
+pub(crate) use start::start_vpn_service;
